==> Notebooks/00_scraping_3_cities.ipynb <==
#   Scrape the neighbours of 3 cities:
  - Berlin
  - Alberta
  - New Orleans

from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.chrome.service import Service as ChromeService
from webdriver_manager.chrome import ChromeDriverManager
import pandas as pd
import numpy as np


driver = webdriver.Chrome(service=ChromeService(ChromeDriverManager().install()))



## Berlin Neighbourhoods

driver.get("https://en.wikipedia.org/wiki/Boroughs_and_neighborhoods_of_Berlin")
Berlin_Neighbourhood=[]
for col in range(12):
    element = driver.find_element(By.XPATH, f'/html/body/div[2]/div/div[3]/main/div[3]/div[3]/div[1]/table[1]/tbody/tr[{col+1}]/td[1]/a').text
    Berlin_Neighbourhood.append(element)

Berlin_df= pd.DataFrame({'Neighbourhood':Berlin_Neighbourhood,'City': 'Berlin', 'Country': 'Germany'})

## Red Deer Alberta Neighbourhoods

driver.get("https://en.wikipedia.org/wiki/Red_Deer,_Alberta#Neighbourhoods")
Alberta_Neighbourhood=[]
for col in range(83):
    element = driver.find_element(By.XPATH, f'/html/body/div[2]/div/div[3]/main/div[3]/div[3]/div[1]/div[4]/ul/li[{col+1}]').text
    Alberta_Neighbourhood.append(element)

Alberta_df= pd.DataFrame({'Neighbourhood':Alberta_Neighbourhood,'City': 'Alberta', 'Country': 'Canada'})

## New Orleans Neighbourhoods

driver.get("https://en.wikipedia.org/wiki/Neighborhoods_in_New_Orleans")
New_orleans_Neighbourhood=[]
for col in range(72):
    element = driver.find_element(By.XPATH, f'/html/body/div[2]/div/div[3]/main/div[3]/div[3]/div[1]/table/tbody/tr[{col+1}]/td[1]').text
    New_orleans_Neighbourhood.append(element)

New_orleans_df= pd.DataFrame({'Neighbourhood':New_orleans_Neighbourhood,'City': 'New Orleans', 'Country': 'United States'})

## Combine 3 Cities into one dataframe

all_df =pd.concat([Berlin_df,Alberta_df,New_orleans_df])

import pickle

all_df.to_pickle('../data/combine_3_cities.pkl')

all_df.to_csv('../data/combine_3_cities.csv',index=False)

driver.close()
