==> Notebooks/01_get_geo_codes.ipynb <==
from geopy.geocoders import Nominatim
import pandas as pd

def get_lat_long(address):
    geolocator = Nominatim(user_agent="Get_Geo_Codes")
    location = geolocator.geocode(address)
    if location is not None:
        lat_ =location.latitude
        long_ = location.longitude
    else:
        lat_ =long_ = None
        
    return lat_,long_
    

all_df = pd.read_csv('../data/combine_3_cities.csv')

lat =[]
long=[]
for index,data in all_df.iterrows():
    address = f'{data["Neighbourhood"]} {data["City"]} {data["Country"]}'
    lat_ , long_ =get_lat_long(address)
    lat.append(lat_)
    long.append(long_)

all_df['Latitude']=lat
all_df['Longitude']=long

## checking Missing Values in Lat and Long

all_df.info()

## Getting some of the missing *values*

for index , data in all_df[(all_df.Latitude.isna())].iterrows():
    address = f'{data["Neighbourhood"]},{data["Country"]}'
    lat_ , long_ =get_lat_long(address)
    all_df.loc[all_df['Neighbourhood']==data['Neighbourhood'],'Latitude']=lat_
    all_df.loc[all_df['Neighbourhood']==data['Neighbourhood'],'Longitude']=long_
    print('The geograpical coordinate of {} are {}, {}.'.format(address, lat_, long_))

all_df.isnull().sum()

for index , data in all_df[(all_df.Latitude.isna())].iterrows():
    address = f'{data["Neighbourhood"]}'
    lat_ , long_ =get_lat_long(address)
    all_df.loc[all_df['Neighbourhood']==data['Neighbourhood'],'Latitude']=lat_
    all_df.loc[all_df['Neighbourhood']==data['Neighbourhood'],'Longitude']=long_
    print('The geograpical coordinate of {} are {}, {}.'.format(address, lat_, long_))

all_df.isnull().sum()

all_df[all_df.Latitude.isna()]

all_df.loc[all_df['Neighbourhood'] == 'Bower Ponds Recreation Area', 'Latitude'] = 52.2688
all_df.loc[all_df['Neighbourhood'] == 'Bower Ponds Recreation Area', 'Longitude'] = -113.8314

all_df.loc[all_df['Neighbourhood'] == 'Gaetz Lakes Sanctuary', 'Latitude'] = 52.2849
all_df.loc[all_df['Neighbourhood'] == 'Gaetz Lakes Sanctuary', 'Longitude'] = -113.7954

all_df.loc[all_df['Neighbourhood'] == 'Morrisroe Extension', 'Latitude'] = 52.2558
all_df.loc[all_df['Neighbourhood'] == 'Morrisroe Extension', 'Longitude'] = -113.7777

all_df.loc[all_df['Neighbourhood'] == 'Red Deer Golf and Country Club', 'Latitude'] = 52.2635
all_df.loc[all_df['Neighbourhood'] == 'Red Deer Golf and Country Club', 'Longitude'] = -113.8400

all_df.loc[all_df['Neighbourhood'] == 'Sunnybrook Extension', 'Latitude'] = 43.7228
all_df.loc[all_df['Neighbourhood'] == 'Sunnybrook Extension', 'Longitude'] = -79.3729

all_df.loc[all_df['Neighbourhood'] == 'Three Mile Bend Recreation Area', 'Latitude'] = 52.3019
all_df.loc[all_df['Neighbourhood'] == 'Three Mile Bend Recreation Area', 'Longitude'] = -113.7817

all_df.loc[all_df['Neighbourhood'] == 'MARLYVILLE - FONTAINEBLEAU', 'Latitude'] = 29.9523
all_df.loc[all_df['Neighbourhood'] == 'MARLYVILLE - FONTAINEBLEAU', 'Longitude'] = -90.1132

all_df.loc[all_df['Neighbourhood'] == 'VIAVANT - VENETIAN ISLES', 'Latitude'] = 30.0178
all_df.loc[all_df['Neighbourhood'] == 'VIAVANT - VENETIAN ISLES', 'Longitude'] = -89.9395

all_df.loc[all_df['Neighbourhood'] == 'TALL TIMBERS - BRECHTEL', 'Latitude'] = 29.9049
all_df.loc[all_df['Neighbourhood'] == 'TALL TIMBERS - BRECHTEL', 'Longitude'] = -89.9976

all_df.loc[all_df['Neighbourhood'] == 'FISCHER DEV', 'Latitude'] = 29.93239931
all_df.loc[all_df['Neighbourhood'] == 'FISCHER DEV', 'Longitude'] = -90.04154205

all_df.loc[all_df['Neighbourhood'] == 'TULANE - GRAVIER', 'Latitude'] = 29.9592
all_df.loc[all_df['Neighbourhood'] == 'TULANE - GRAVIER', 'Longitude'] = -90.0849

all_df.loc[all_df['Neighbourhood'] == 'NEW AURORA - ENGLISH TURN', 'Latitude'] = 44.8408158
all_df.loc[all_df['Neighbourhood'] == 'NEW AURORA - ENGLISH TURN', 'Longitude'] = -93.5191807

all_df.loc[all_df['Neighbourhood'] == 'Riverside Light Industrial Park', 'Latitude'] = 25.04456915
all_df.loc[all_df['Neighbourhood'] == 'Riverside Light Industrial Park', 'Longitude'] = -77.3314679919205

all_df.loc[all_df['Neighbourhood'] == 'McKenzie Trail Recreation Area', 'Latitude'] = 48.1047453
all_df.loc[all_df['Neighbourhood'] == 'McKenzie Trail Recreation Area', 'Longitude'] = -103.1035193


nan = all_df[all_df.Latitude.isna()]

nan.index

all_df.drop(nan.index,inplace=True)

all_df[all_df.Latitude.isna()]

all_df

all_df.to_csv('../data/df_with_lat_long.csv',index=False)
