==> Notebooks/03_getting_venues.ipynb <==
CLIENT_ID = '2GQBW5PR0QFXTOGCHKTRFWJBTGOFOHXW1TRTNRAFURQ5FE1X'
CLIENT_SECRET = '3QH40WMZIIDSQN1RFAVAEQHUIMOQUJPKYPABQVNTSDQJN2YD'
VENU_KEY='fsq3ZvwbThl+Fk13tHHRPsLUDh5Mp1mxVvNVGwdzaCyOJh4='
VERSION = 20202808
LIMIT = 100

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import requests, json

def getNearbyVenues(names, latitudes, longitudes, radius=1000): #names: City Neighbourhoods
    venues_list=[]
    for name, lat, lng in zip(names, latitudes, longitudes):
        print(name)  #Neighbourhoods name
            
        # create the API request URL
        url = 'https://api.foursquare.com/v2/venues/explore?&client_id={}&client_secret={}&v={}&ll={},{}&radius={}&limit={}'.format(
            CLIENT_ID, 
            CLIENT_SECRET, 
            VERSION, 
            lat, 
            lng, 
            radius, 
            LIMIT)
        
        # make the GET request
        results = requests.get(url).json()["response"]['groups'][0]['items']
        if (len(results)==0):
            print('0')
        # return only relevant information for each nearby venue
        venues_list.append([(
            name, 
            lat, 
            lng, 
            v['venue']['name'], 
            v['venue']['location']['lat'], 
            v['venue']['location']['lng'],  
            v['venue']['categories'][0]['name']) for v in results]) #v loops over items "venues"

    #print("Venues list ", venues_list)   #list of Venues for all Neighbourhoods of the City
    print("Number of Neighbourhoods at the City ", len(venues_list))

    nearby_venues = pd.DataFrame([venue for neigh in venues_list for venue in neigh])
    # venues_list : list of lists of tubles "list of neighbourhoods, of list of venues, of tuble of venue data"
    nearby_venues.columns = ['Neighbourhood',
                             'Neighbourhood Latitude',
                             'Neighbourhood Longitude',
                             'Venue',
                             'Venue Latitude',
                             'Venue Longitude',
                             'Venue Category']

    return(nearby_venues)

df_neigh = pd.read_csv("../data/final_df.csv")
df_neigh

df_berlin_venues = getNearbyVenues(df_neigh.loc[df_neigh['City']=='Berlin', 'Neighbourhood'],
                               df_neigh.loc[df_neigh['City']=='Berlin', 'Latitude'],
                               df_neigh.loc[df_neigh['City']=='Berlin', 'Longitude'], 1000)

df_berlin_venues

df_alberta_venues = getNearbyVenues(df_neigh.loc[df_neigh['City']=='Alberta', 'Neighbourhood'],
                               df_neigh.loc[df_neigh['City']=='Alberta', 'Latitude'],
                               df_neigh.loc[df_neigh['City']=='Alberta', 'Longitude'], 1000)

df_alberta_venues

df_newOrleans_venues = getNearbyVenues(df_neigh.loc[df_neigh['City']=='New Orleans', 'Neighbourhood'],
                               df_neigh.loc[df_neigh['City']=='New Orleans', 'Latitude'],
                               df_neigh.loc[df_neigh['City']=='New Orleans', 'Longitude'], 1000)

df_newOrleans_venues

df_newOrleans_venues.to_csv('../data/df_newOrleans_venues.csv',index=False)
df_alberta_venues.to_csv('../data/df_alberta_venues.csv',index=False)
df_berlin_venues.to_csv('../data/df_berlin_venues.csv',index=False)
